+++ retail/modules/nameFix.lua
@@ -1006,7 +1006,7 @@ function BBF.SetCustomFonts()
             SystemFont_Outline_WTF2,
             GameTooltipHeader,
             System_IME,
-            Number12Font_o1
+            Number12Font_o1,
         }
 
         -- Backup function for the chat font